extract WithContentHelper
